Add support for grouping sequences into directories
Determine number of images in sequence based on length of sequence.
Increase minimum number of images in a sequence.
Set a minimum interval between images to be considered a time-lapse.
This helps filter continuous shooting sequences in normal photography.
Add check to see if any files match the given pattern.